Remove default_filter module, put into logging
